+++ resources/ReleaseNotes.md
@@ -1,6 +1,7 @@
 # Software
 
 - Export I-ALiRT data
+- Add operations to plot composite variables (e.g., difference between variables)
 - Add optional global legend to figures generated with `mag.graphics.visualize`
 - Add option to change tile indexing in figures generated with `mag.graphics.visualize`
 - Fix wrong values for I-ALiRT data and packet frequency